requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
backports.tarfile==1.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
colorlog==6.8.2
comm==0.2.2
confection==0.1.5
cymem==2.0.8
debugpy==1.8.5
decorator==5.1.1
dnspython==2.6.1
docutils==0.21.1
docx2pdf==0.1.8
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
executing==2.1.0
fastapi==0.110.1
h11==0.14.0
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.5
ipython==8.27.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.1.0
langcodes==3.4.0
language_data==1.2.0
lxml==5.2.1
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.2.0
murmurhash==1.0.10
nest-asyncio==1.6.0
nh3==0.2.17
nltk==3.9.1
numpy==1.26.4
openpyxl==3.1.2
packaging==24.1
parso==0.8.4
pdf2image==1.17.0
pillow==10.4.0
pkginfo==1.10.0
plac==1.4.3
platformdirs==4.3.6
plum-dispatch==1.7.4
preshed==3.0.9
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
pymongo==4.9.1
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-multipart==0.0.9
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==26.2.0
readme_renderer==43.0
regex==2024.9.11
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
setuptools==69.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Spire.Doc==12.3.2
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
thinc==8.2.5
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
twine==5.0.0
typer==0.12.5
typing_extensions==4.11.0
ultra_logger==0.1.2
urllib3==2.2.1
uvicorn==0.29.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wheel==0.43.0
wrapt==1.16.0
zipp==3.18.1
